+++ inst/shiny/DiagnosticsExplorer/server.R
@@ -34,8 +34,49 @@ shinyServer(function(input, output, session) {
   
   # Tables
   
-  output$incidenceTable <- renderDT({
-    table <- incidence
+  output$proportionTable <- renderDT({
+    
+    analysisId <- settings %>% filter(cdmdatabasename == input$pdatabase & 
+                                        tar == input$ptar &
+                                        outcomeid == getId(input$poutcomeName) &
+                                        cohortid ==  getId(input$ptargetName)
+    ) %>% select(analysisid)
+    analysisId <-analysisId$analysisid[1]
+    
+    if(input$pxyear){
+     proportionAll <- proportion %>% dplyr::filter(agegroup == 'all')
+    }else{
+      proportionAll <- proportion %>% dplyr::filter(year == 'all') 
+    }
+    
+    table <- proportionAll %>% filter(analysisid == analysisId) %>%
+      select(year,agegroup,gender,n,o,opercent)
+    
+    if(input$pgender){
+      table <- table %>% filter(gender != -1)
+    }else{
+      table <- table %>% filter(opercent != -1) %>% 
+        group_by(year,agegroup) %>%
+        summarise(n = sum(as.double(as.character(n)), na.rm = T),
+                  o = sum(as.double(as.character(o)), na.rm = T),
+                  opercent = sum(o, na.rm = T)*100/sum(n, na.rm = T))
+    }
+    
+    # get the correct columns
+    if(input$pxyear & input$pgender){
+      table <-table %>% ungroup() %>%
+        select(year,gender,n,o,opercent)
+    } else if(!input$pxyear & input$pgender){
+      table <-table %>% ungroup() %>%
+        select(agegroup,gender,n,o,opercent)
+    } else if(!input$pxyear & !input$pgender){
+      table <-table %>% ungroup() %>%
+        select(agegroup,n,o,opercent)
+    } else if(input$pxyear & !input$pgender){
+      table <- table %>% ungroup() %>%
+        select(year,n,o,opercent)
+    }
+    
     table$opercent = as.numeric(prettyHr(table$opercent))
     result<-datatable(table,
       filter="top",
@@ -49,6 +90,158 @@ shinyServer(function(input, output, session) {
       class = "stripe nowrap compact")
     return(result)}
   )
+  
+  output$proportionPlot <- renderPlotly({
+    mySubplot <- function(myanalysisId,myDatabasename, myTar){
+      
+      if(input$pxyear){
+        proportionAll <- proportion %>% dplyr::filter(agegroup == 'all')
+      }else{
+        proportionAll <- proportion %>% dplyr::filter(year == 'all') 
+      }
+      
+      plotdata <- proportionAll %>% filter(analysisid == myanalysisId)
+      plotdataM <- plotdata %>% filter(gender == 8507)
+      plotdataF <- plotdata %>% filter(gender == 8532)
+      plotdataA <- plotdata %>% filter(opercent != -1) %>% 
+        group_by(year,agegroup) %>%
+        summarise(n = sum(as.double(as.character(n)), na.rm = T),
+                  o = sum(as.double(as.character(o)), na.rm = T),
+                  opercent = sum(o, na.rm = T)*100/sum(n, na.rm = T)) %>%
+        ungroup() %>% data.frame()
+      
+      fig1 <- plot_ly()
+      if(input$pxyear){
+        levels <- c(2005:2020)
+      } else{
+        levels <- c('0 - 4','5 - 9','10 - 14','15 - 19', '20 - 24',
+                    '25 - 29','30 - 34','35 - 39','40 - 44','45 - 49',
+                    '50 - 54','55 - 59','60 - 64','65 - 69','70 - 74',
+                    '75 - 79','80 - 84','85 - 89','90 - 94','95 - 99')
+      }
+      plotdataF$var <- factor(as.character(plotdataF[,ifelse(input$pxyear,'year','agegroup')]), 
+                              levels = levels,ordered = TRUE)
+      plotdataM$var <- factor(as.character(plotdataM[,ifelse(input$pxyear,'year','agegroup')]), 
+                                           levels = levels, ordered = TRUE)
+      plotdataA$var <- factor(as.character(plotdataA[,ifelse(input$pxyear,'year','agegroup')]), 
+                                           levels = levels, ordered = TRUE)
+      plotdataF <- plotdataF[order(plotdataF$var),]
+      plotdataM <- plotdataM[order(plotdataM$var),]
+      plotdataA <- plotdataA[order(plotdataA$var),]
+      
+      if(input$pgender){
+      fig1 <- fig1 %>%
+        add_trace(
+          x = plotdataF$var,
+          y = plotdataF$opercent,
+          type = 'scatter',
+          name = '',
+          mode = 'lines+markers',
+          showlegend = FALSE,
+          hovertemplate = paste('<b>',ifelse(input$pxyear,'year','agegroup'),'</b>: ',plotdataF$var,
+                                '<br><i>T size </i>',plotdataF$n,'<br>',
+                                '<i>O percent</i>: %{y:.2f}',
+                                '<br>')
+        )
+      
+      fig1 <- fig1 %>%
+        add_trace(
+          x = plotdataM$var,
+          y = plotdataM$opercent,
+          type = 'scatter',
+          name = '',
+          mode = 'lines+markers',
+          showlegend = FALSE, 
+          hovertemplate = paste('<b>',ifelse(input$pxyear,'year','agegroup'),'</b>: ',plotdataM$var,
+                                '<br><i>T size </i>',plotdataM$n,'<br>',
+                                '<i>O percent</i>: %{y:.2f}',
+                                '<br>')
+          #markers=list(color='red',dash='dashed')
+        )
+      } else{
+        fig1 <- fig1 %>%
+          add_trace(
+            x = plotdataA$var,
+            y = plotdataA$opercent,
+            type = 'scatter',
+            name = '',
+            mode = 'lines+markers',
+            showlegend = FALSE, 
+            hovertemplate = paste('<b>',ifelse(input$pxyear,'year','agegroup'),'</b>: ',plotdataA$var,
+                                  '<br><i>T size </i>',plotdataA$n,'<br>',
+                                  '<i>O percent</i>: %{y:.2f}',
+                                  '<br>')
+          )
+      }
+      
+      fig1
+    }
+    
+    subplots_list <- list()
+    ind <- 1
+    for(i in 1:length(input$databases)){
+      
+      analysisId <- settings %>% filter(cdmdatabasename == input$pdatabases[i] & 
+                                          tar == input$ptar &
+                                          outcomeid == getId(input$poutcomeName) &
+                                          cohortid ==  getId(input$ptargetName)
+      ) %>% select(analysisid)
+      analysisId <-analysisId$analysisid[1]
+      
+      # if there is a result add it 
+      if(!is.na(analysisId)){
+        tempPlot <- mySubplot(analysisId, input$databases[i], input$ptar)
+        subplots_list[[ind]] <- tempPlot
+        ind <- ind+1
+      }
+    }
+    
+    
+    
+    fig <- subplot(subplots_list, nrows=length(input$databases),shareX = F, shareY =F)
+    
+    return(fig)
+  })
+  
+  
+  output$survivalPlot <- renderPlot({
+    
+    data <- survival %>% 
+      dplyr::filter(outcomeid == getId(input$soutcomeName)) %>%
+      dplyr::filter(cohortid == getId(input$stargetName)) %>%
+      dplyr::filter(cdmdatabasename == input$sdatabase)
+    
+    data <- data %>% dplyr::mutate(decreaseP = events/(events+natrisk))
+    
+    yaxis <- lapply(unique(data$censoredtime), function(i) 1-sum(data %>% dplyr::filter(censoredtime <= i) %>% dplyr::select(decreaseP)))
+    
+    extra <- data.frame(censoredtime = unique(data$censoredtime),
+                        yaxis = unlist(yaxis))
+    
+    data <- data %>% inner_join(extra, by = 'censoredtime')
+    
+    library(ggplot2)
+    ggplot() +
+      geom_step(data=data, mapping=aes(x=censoredtime, y=yaxis)) +
+      #geom_step(data=d, mapping=aes(x=x, y=y), direction="vh", linetype=3) +
+      geom_point(data=data, mapping=aes(x=censoredtime, y=yaxis), color="red") +
+      geom_vline(xintercept = 365, linetype="dotted", 
+                 color = "black", size=1) +
+      geom_vline(xintercept = 2*365, linetype="dotted", 
+                 color = "black", size=1) +
+      geom_vline(xintercept = 3*365, linetype="dotted", 
+                 color = "black", size=1) + 
+      geom_vline(xintercept = 4*365, linetype="dotted",
+                 color = "black", size=1) +
+      geom_vline(xintercept = 5*365, linetype="dotted", 
+                 color = "black", size=1) +
+      geom_vline(xintercept = 10*365, linetype="dotted", 
+                 color = "black", size=1) +
+      ylim(0, 1) + labs(x = "Time from index (days)",
+                        y = "Outcome free")
+    
+  })
+
   
   output$characterizationTable <- renderDT({