+++ NEWS.md
@@ -1,3 +1,12 @@
+PatientLevelPrediction 4.1.0
+======================
+- added updates and fixes into master from development branch
+
+PatientLevelPrediction 4.0.6
+======================
+- fixed bug with pdw data extraction due to multiple person_id columns
+- fixed bug in shiny app converting covariate values due to tibble
+
 PatientLevelPrediction 4.0.5
 ======================
 - added calibration updates: cal-in-large, weak cal